Change int to byte where possible.

+++ src/de/gwvprojekt/algorithm/AStar.java
@@ -13,7 +13,7 @@ public class AStar extends AbstractAlgorithm{
 		State goal = start.goalState();
 		
 		PriorityQueue<Node> frontier = new PriorityQueue<Node>();
-		frontier.add(new NodeImpl(heuristic.calculateFor(start), start, null, 0));
+		frontier.add(new NodeImpl(heuristic.calculateFor(start), start, null, (byte) 0));
 		
 		Collection<State> closed = new HashSet<State>();
 		
@@ -27,7 +27,7 @@ public class AStar extends AbstractAlgorithm{
 			}
 			
 			currentPathCost++;
-			for(int direction : current.getState().getPossibleDirections()){
+			for(byte direction : current.getState().getPossibleDirections()){
 				State neighbor = current.getState().move(direction);
 				if(!closed.contains(neighbor)){
 					frontier.add(new NodeImpl(currentPathCost + heuristic.calculateFor(neighbor),
