+++ src/de/gwvprojekt/algorithm/Node.java
@@ -4,5 +4,5 @@ public interface Node extends Comparable<Node>{
 	public int   getEstimatedCost();
 	public State getState();
 	public Node  getPredecessor();
-	public int   getDirection();
+	public byte  getDirection();
 }
