+++ src/de/gwvprojekt/algorithm/State.java
@@ -4,9 +4,9 @@ import java.util.Random;
 
 public interface State
 {
-	public abstract int   getValue(int row, int column);
-	public abstract int[] getPos(int value);
-	public abstract int[] getPossibleDirections();
+	public abstract byte   getValue(byte row, byte column);
+	public abstract byte[] getPos(byte value);
+	public abstract byte[] getPossibleDirections();
 	
 	public void randomizeData(Random random);
 	
@@ -16,16 +16,16 @@ public interface State
 	 * @param int row, int column
 	 * @return string
 	 */
-	public String getStringValue(int row, int column);
+	public String getStringValue(byte row, byte column);
 
-	public State move(int direction) throws CloneNotSupportedException; // 1 -> up, 2 -> left, 3 -> right, 4 -> down
+	public State move(byte direction) throws CloneNotSupportedException; // 1 -> up, 2 -> left, 3 -> right, 4 -> down
 
 	/**
 	 * vertauscht das leere Feld mit dem angegebenen Feld, falls dies möglich ist
 	 * @param row
 	 * @param column
 	 */
-	public void move(int row, int column);
+	public void move(byte row, byte column);
 
 	/**
 	 * Prüft ob Tauschen möglich ist
@@ -33,7 +33,7 @@ public interface State
 	 * @param column
 	 * @return
 	 */
-	public boolean isMovable(int row, int column);
+	public boolean isMovable(byte row, byte column);
 
 	/**
 	 * Initialisiert die Matrix mit einer Permutation
@@ -42,5 +42,5 @@ public interface State
 	
 	public State goalState();
 	
-	public int getNumberOfMisplacedTiles();
+	public byte getNumberOfMisplacedTiles();
 }
